+++ source/constructs/test/live-streaming.test.ts
@@ -25,8 +25,8 @@ expect.addSnapshotSerializer({
                 replacementValue: 'AssetParameters[HASH REMOVED]'
             },
             {
-                regex: /(\w+ for asset)\s?(version)?\s?"([A-Fa-f0-9]{64})"/,
-                replacementValue: '$1 [HASH REMOVED]'
+                regex: /^[A-Fa-f0-9]{64}(\.zip)$/,
+                replacementValue: '[HASH REMOVED]$1'
             }
         ];